+++ shared/js/background/dnr-session-rule-id.js
@@ -8,10 +8,6 @@ const SESSION_RULE_STORAGE_KEY = 'sessionRuleOffset'
 let sessionRuleOffset = 0
 let ready = false
 
-export function isValidSessionId (id) {
-    return id >= SESSION_RULE_ID_START
-}
-
 export async function setSessionRuleOffsetFromStorage () {
     const offset = await browserWrapper.getFromSessionStorage(SESSION_RULE_STORAGE_KEY)
     if (offset) {
@@ -34,4 +30,25 @@ export function getNextSessionRuleId () {
     sessionRuleOffset += 1
     browserWrapper.setToSessionStorage(SESSION_RULE_STORAGE_KEY, sessionRuleOffset)
     return nextRuleId
+}
+
+function isValidSessionId (id) {
+    return id >= SESSION_RULE_ID_START
+}
+
+/**
+* Remove orphaned session ids
+* We increment the rule IDs for some session rules, starting at STARTING_RULE_ID and
+* keep a note of the next rule ID in session storage. During extesion update/restarts
+* session storage is cleared, while session rules are not, which causes errors due to
+* session rule ID conflicts
+ * @return {Promise}
+ */
+export function flushSessionRules () {
+    return chrome.declarativeNetRequest.getSessionRules().then(rules => {
+        const ruleIds = rules.map(({ id }) => id).filter(isValidSessionId)
+        if (ruleIds.length) {
+            return chrome.declarativeNetRequest.updateSessionRules({ removeRuleIds: ruleIds })
+        }
+    })
 }
